Reconcile `scalems.rp` and `scalems.radical`.
* Recombine and rearrange. Move new function dispatching tools from top
  level to `scalems.radical.runtime` submodule.
* Introduce a configuration option to enable/disable the `rp.raptor`
  management.
